Pharmatrix-main/medi/backend/requirements.txt
twilio
python-dotenv
fastapi
uvicorn
sqlalchemy
python-multipart
requests
pillow
pytesseract
passlib[bcrypt]
pymongo[srv]
dnspython
bcrypt
